Rename ctorData to ctorConfig
Summary:

Previously, ctorData is accessing all the attributes listed in ErrorConfig so it should be define as ErrorConfig instance and renamed as ctorConfig.

The common use case will bind the first 2 arguments of ApolloError.constructor() to name and a common configuration object, leaving the ctorConfig to provide the final customization.

We make sure that ctorConfig take precedence over config and merge them whenever necessary.